Port WIP chunk file reader (.cchk_pc & .gchk_pc files)
The original was incomplete and could only read some of the chunk files. It had some hacky code to try and fix those issues, but that still wasn't enough.
This time I ported the code that I know works, and left the hacky code out. I want to fix the problems with extracting chunk meshes for good before I start using this in the Nanoforge rewrite.